Allow user invites with env mailer config
